feat: add support for component router
