+++ src/backend/adapters/angular2.ts
@@ -25,9 +25,7 @@ import { Subject } from 'rxjs/Subject';
 
 import { TreeNode, BaseAdapter } from './base';
 import { Description } from '../utils/description';
-import { ParseRouter, ComponentRouterParser } from '../utils/parse-router';
-
-import { IS_OLD_ROUTER } from './router-checker';
+import { ParseRouter, IS_OLD_ROUTER_HACK, parseConfigRoutes } from '../utils/parse-router';
 
 export class Angular2Adapter extends BaseAdapter {
   _tree: any = {};
@@ -75,11 +73,11 @@ export class Angular2Adapter extends BaseAdapter {
       const router = ng.probe(root).componentInstance.router;
       let routes: any;
 
-      if (IS_OLD_ROUTER(router)) {
-        // TODO: (ericjim): remove if block and function once we no longer support the old router.
+      if (IS_OLD_ROUTER_HACK(router)) {
+        // TODO: (ericjim): remove if-block and function once we no longer support the old router.
         routes = ParseRouter.parseRoutes(router.root.registry);
       } else {
-        routes = ComponentRouterParser.parseConfigRoutes(router);
+        routes = parseConfigRoutes(router);
       }
 
       this.showRoutes(routes);
