+++ src/frontend/components/router-tree/router-tree.ts
@@ -44,7 +44,6 @@ export class RouterTree {
   }
 
   render() {
-    console.log(this.routerTree);
     if (!this.routerTree) {
       return;
     }
